chore: rename and separate heating-mat classes & fix minor implimentation
